The font type can now be selected in the display web page. No need to change the slot configuration JSON file manually anymore.

+++ lib/Views/src/layout/DateTimeView32x16.h
@@ -341,63 +341,63 @@ public:
 protected:
 
     /** Distance between two lamps in pixel. */
-    static const uint8_t LAMP_DISTANCE = 1U;
+    static const uint16_t LAMP_DISTANCE = 1U;
 
     /** Lamp width in pixel. */
-    static const uint8_t LAMP_WIDTH    = (CONFIG_LED_MATRIX_WIDTH - ((MAX_LAMPS + 1U) * LAMP_DISTANCE)) / MAX_LAMPS;
+    static const uint16_t LAMP_WIDTH    = (CONFIG_LED_MATRIX_WIDTH - ((MAX_LAMPS + 1U) * LAMP_DISTANCE)) / MAX_LAMPS;
 
     /** Lamp distance to the canvas border in pixel. */
-    static const uint8_t LAMP_BORDER   = (CONFIG_LED_MATRIX_WIDTH - (MAX_LAMPS * LAMP_WIDTH) - ((MAX_LAMPS - 1U) * LAMP_DISTANCE)) / 2U;
+    static const uint16_t LAMP_BORDER   = (CONFIG_LED_MATRIX_WIDTH - (MAX_LAMPS * LAMP_WIDTH) - ((MAX_LAMPS - 1U) * LAMP_DISTANCE)) / 2U;
 
     /** Lamp height in pixel. */
-    static const uint8_t LAMP_HEIGHT   = 1U;
+    static const uint16_t LAMP_HEIGHT   = 1U;
 
     /** Lamp 0 x-coordinate in pixel. */
-    static const uint8_t LAMP_0_X      = LAMP_BORDER + (0 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_0_X       = LAMP_BORDER + (0 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 1 x-coordinate in pixel. */
-    static const uint8_t LAMP_1_X      = LAMP_BORDER + (1 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_1_X       = LAMP_BORDER + (1 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 2 x-coordinate in pixel. */
-    static const uint8_t LAMP_2_X      = LAMP_BORDER + (2 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_2_X       = LAMP_BORDER + (2 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 3 x-coordinate in pixel. */
-    static const uint8_t LAMP_3_X      = LAMP_BORDER + (3 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_3_X       = LAMP_BORDER + (3 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 4 x-coordinate in pixel. */
-    static const uint8_t LAMP_4_X      = LAMP_BORDER + (4 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_4_X       = LAMP_BORDER + (4 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 5 x-coordinate in pixel. */
-    static const uint8_t LAMP_5_X      = LAMP_BORDER + (5 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_5_X       = LAMP_BORDER + (5 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp 6 x-coordinate in pixel. */
-    static const uint8_t LAMP_6_X      = LAMP_BORDER + (6 * (LAMP_WIDTH + LAMP_DISTANCE));
+    static const int16_t LAMP_6_X       = LAMP_BORDER + (6 * (LAMP_WIDTH + LAMP_DISTANCE));
 
     /** Lamp y-coordindate in pixel. */
-    static const uint8_t LAMP_Y        = CONFIG_LED_MATRIX_HEIGHT - 1U;
+    static const int16_t LAMP_Y         = CONFIG_LED_MATRIX_HEIGHT - 1U;
 
     /**
      * Text width in pixels.
      */
-    static const uint16_t TEXT_WIDTH   = CONFIG_LED_MATRIX_WIDTH;
+    static const uint16_t TEXT_WIDTH    = CONFIG_LED_MATRIX_WIDTH;
 
     /**
      * Text height in pixels.
      */
-    static const uint16_t TEXT_HEIGHT  = CONFIG_LED_MATRIX_HEIGHT - LAMP_HEIGHT;
+    static const uint16_t TEXT_HEIGHT   = CONFIG_LED_MATRIX_HEIGHT - LAMP_HEIGHT;
 
     /**
      * Text widget x-coordinate in pixels.
      */
-    static const int16_t TEXT_X        = 0;
+    static const int16_t TEXT_X         = 0;
 
     /**
      * Text widget y-coordinate in pixels.
      */
-    static const int16_t TEXT_Y        = 0;
+    static const int16_t TEXT_Y         = 0;
 
     /** Start of week offset for the week bar (Sunday = 0). */
-    static const uint8_t START_OF_WEEK = 1U;
+    static const uint8_t START_OF_WEEK  = 1U;
 
     /** Color of the current day shown in the day of the week bar. */
     static const Color DAY_ON_COLOR;
